aircraft-json-to-bucket/requirements.txt
requests==2.21.0
google-cloud-storage==1.14.0
nose==1.3.7